Completely Finished Combo Box
The combo box is completely done, including some additional features
that are helpful, and make usage smoother including, when scrolling, the
highlighted item is automatically updated;  when using the up/down keys,
if the reach either the top or bottom, the respective one disables; the
button of the combobox no longer overlaps with the up arrow.  Major
Accomplishments are: getting scroll bar to work perfectly, and finishing
keyboard input for the combo box.