Initial work on List Box
I started and got major work done on the list box, most of it is done,
but the multiple selection does NOT work yet.  The scrolling works, and
selection works too.  I changed the system to work similarly to how the
combo box works, but it is a little bit simpler.  I also fixed the font
coloring issue.