Initial work on Edit box
I have started work on the edit box, and the UniBuffer as well.  I am
getting a strange unhandled exception when calling GLUFOffsetRect, but I
still have to investigate that.  I do not know if it is an exception at
the memory of the function, or the memory of the rect being passed.