Edit box with new system
I finished the edit box analysis system, so we have bounding boxes for
each character, but the next thing to do is to make the box editable,
which will require the hardest step, which is translating render space
characters, to characters in the actual string and back.  But before I
do that, I need to get the cursor blink to work, because that will make
debugging a lot easier.