New Idea
I had the idea of having two different Vertex Array classes, one for
AoS, and one for SoA.   I have finished the idea, but I have to perfect
it by testing it in a real world situation: loading a file from assimp,
and loading it in.  It is probably going to be more efficient to use SoA
for models, and AoS for manually entered data.