Back to pixel coords, and no GLI
GLI was not cooperating with creating textures from memory, so I ditched
that, and i decided to go back to pixel coords, because that seems to be
simpler.  The only thing I need to do is fix a bug where the control
textures will not render.  I will also add further function to the edit
box, such as single-line (auto-scroll), but also see if that still works
when in pixel space.  I will finish the edit box, because it is now
incomplete.