Tons of Edit Box Work
The formatting works much better on the multiline box now, but I still
have to completely fix the single line one, where it is not scrolling
quite far enough over.  I the multiline one there is a bug that causes
an assertion error when scrolling, and I need to fix that.